requirements.txt
numpy == 1.24.4
torch == 2.1.2
torchvision == 0.16.2
streamlit == 1.29.0
pillow == 10.2.0
-e .